fix(arc-1033): lazy load flowplayer => serverside rendered in nextjs
https://meemoo.atlassian.net/browse/ARC-1033
